Move the garbage collection code from eval.c into a separate gc.c file
